Registry: implement scheduled refreshing of registry data
