+++ .env
@@ -28,6 +28,7 @@ APP_SECRET=8a8032bdfb2faee86d25803ec1a98f1c
 DATABASE_URL="postgresql://app:!ChangeMe!@database:5432/app?serverVersion=15&charset=utf8"
 ###< doctrine/doctrine-bundle ###
 
+REGISTRY_DATABASE_REFRESH_ENABLED=true
 REGISTRY_DATABASE_CH_FILE="%kernel.project_dir%/var/data/data-ch.sqlite"
 REGISTRY_DATABASE_LI_FILE="%kernel.project_dir%/var/data/data-li.csv"
 MEILISEARCH_DSN="http://meilisearch:7700?apiKey=cWwx3jxAworhaQAVHeRYlK37wdVnLNc6P_Uyz8WLTyM"
