+++ composer.json
@@ -20,6 +20,7 @@
         "doctrine/doctrine-bundle": "^2.11",
         "doctrine/doctrine-migrations-bundle": "^3.3",
         "doctrine/orm": "^2.17",
+        "dragonmantank/cron-expression": "^3.4",
         "jsor/doctrine-postgis": "dev-dbal-deprecations#c82761b99023f09e2b9f5ea53baf8719ee7a5b02 as v2.2",
         "league/csv": "^9.22",
         "meilisearch/meilisearch-php": "^1.13",
